Using dat is unnecessary. Let's convert to JSON structure
