+++ README.md
@@ -1,5 +1,3 @@
 # Hi how's it going idk what I'm doing yet
 
-# Here's a line added on the remote repo that I will be pulling on my local repo
-
-# Ohhh so the hash symbol is header syntax, I see
+My first multi-file project idea. I took a detour to work on my Stock Manager project before completing it. With what I've learned from that experience, I'm going to be working on restructuring and finishing this!
