+++ settings.py
@@ -1,5 +1,5 @@
 import date_logic
-import file_modification
+import file_management
 
 
 def set_week_day(line_data, input_week_day):
